Support non-Array messages for BarbequeClient::Executor
ActiveJob expects only Array for its argments. Therefore if other class'
instance cames, it crashes or works with unintended behavior. Especially
Hash (e.g. `{ "foo" => "bar" }`) cames, Job's #perform method receives
a key-value pair Array (e.g. `["foo", "bar"]`) as an argument. It's
unexpected and not useful for most cases. So with this patch if the
argument is not Array, it's transformed into one element Array with the
original argument (`{ "foo" => "bar" }` to `[{ "foo" => "bar" }]`.